pyproject.toml
[project]
name = "ask"
version = "0.1.0"
description = "A minimal openrouter cli."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
